Change how hoverTextTables are built/displayed
position: fixed, and only one instance for each user. Jumps to the
correct position, and avoids letting the bottom go off screen.

This is a very different way to display the hover text, and there are
already some small bugs that I'm seeing, such as maybe a 1% chance that
the hovertext doesn't go away correctly without mousing over it again.
Also the fact that it moves as the screen scrolls is sort of ugly.