Check bid api for faster updates
All in all, now a change in rankings should be represented within about
30 seconds of the trade being made.

In the background.js, Full updates happen less regularly, and now there
are partial updates, which check the bid api. If there are new bids to
be found there, checks if any of them are on the relevant markets. If
so, fetch and update just those market positions.

On the fontend, occasionally do a clean rerun of the hover table
generation to pick up changes.

This commit probably introduces a lot of bugs.