Make two partial update attempts after bets
At 10 seconds and 30 seconds. It's important that we don't give up too
early, otherwise the change will not be picked up until a full sync or
the next bet on that market. Sadly, I can't think of a good way to
verify that the change represented by the bet has been accounted for
based on the bet api.